% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Build}
\alias{Build}
\title{Build Object}
\usage{
Build(Build.substitutions = NULL, Build.timing = NULL, results = NULL,
  logsBucket = NULL, steps = NULL, buildTriggerId = NULL, id = NULL,
  tags = NULL, startTime = NULL, substitutions = NULL, timing = NULL,
  createTime = NULL, sourceProvenance = NULL, images = NULL,
  projectId = NULL, logUrl = NULL, finishTime = NULL, options = NULL,
  source = NULL, timeout = NULL, status = NULL, statusDetail = NULL,
  secrets = NULL)
}
\arguments{
\item{Build.substitutions}{The \link{Build.substitutions} object or list of objects}

\item{Build.timing}{The \link{Build.timing} object or list of objects}

\item{results}{Results of the build}

\item{logsBucket}{Google Cloud Storage bucket where logs should be written (see}

\item{steps}{Required}

\item{buildTriggerId}{The ID of the `BuildTrigger` that triggered this build, if it was}

\item{id}{Unique identifier of the build}

\item{tags}{Tags for annotation of a `Build`}

\item{startTime}{Time at which execution of the build was started}

\item{substitutions}{Substitutions data for `Build` resource}

\item{timing}{Stores timing information for phases of the build}

\item{createTime}{Time at which the request to create the build was received}

\item{sourceProvenance}{A permanent fixed identifier for source}

\item{images}{A list of images to be pushed upon the successful completion of all build}

\item{projectId}{ID of the project}

\item{logUrl}{URL to logs for this build in Google Cloud Console}

\item{finishTime}{Time at which execution of the build was finished}

\item{options}{Special options for this build}

\item{source}{The location of the source files to build}

\item{timeout}{Amount of time that this build should be allowed to run, to second}

\item{status}{Status of the build}

\item{statusDetail}{Customer-readable message about the current status}

\item{secrets}{Secrets to decrypt using Cloud Key Management Service}
}
\value{
Build object
}
\description{
Build Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A build resource in the Container Builder API.At a high level, a `Build` describes where to find source code, how to buildit (for example, the builder image to run on the source), and where to storethe built artifacts.Fields can include the following variables, which will be expanded when thebuild is created:- $PROJECT_ID: the project ID of the build.- $BUILD_ID: the autogenerated ID of the build.- $REPO_NAME: the source repository name specified by RepoSource.- $BRANCH_NAME: the branch name specified by RepoSource.- $TAG_NAME: the tag name specified by RepoSource.- $REVISION_ID or $COMMIT_SHA: the commit SHA specified by RepoSource or  resolved from the specified branch or tag.- $SHORT_SHA: first 7 characters of $REVISION_ID or $COMMIT_SHA.
}
\seealso{
Other Build functions: \code{\link{Build.substitutions}},
  \code{\link{Build.timing}},
  \code{\link{projects.builds.create}}
}
