% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Build.timing}
\alias{Build.timing}
\title{Build.timing Object}
\usage{
Build.timing()
}
\value{
Build.timing object
}
\description{
Build.timing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Stores timing information for phases of the build. Valid keys are:* BUILD: time to execute all build steps* PUSH: time to push all specified images.* FETCHSOURCE: time to fetch source.If the build does not specify source or images,these keys will not be included.@OutputOnly
}
\seealso{
Other Build functions: \code{\link{Build.substitutions}},
  \code{\link{Build}}, \code{\link{projects.builds.create}}
}
