% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildOperationMetadata}
\alias{BuildOperationMetadata}
\title{Cloud Container Builder API Objects 
Builds container images in the cloud.}
\usage{
BuildOperationMetadata(build = NULL)
}
\arguments{
\item{build}{The build that the operation is tracking}
}
\value{
BuildOperationMetadata object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2018-03-22 19:45:45
filename: ~/Desktop/googlecontainer/googlecloudbuildv1.auto/R/cloudbuild_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
BuildOperationMetadata Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for build operations.
}
