% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildOptions}
\alias{BuildOptions}
\title{BuildOptions Object}
\usage{
BuildOptions(logStreamingOption = NULL, requestedVerifyOption = NULL,
  substitutionOption = NULL, diskSizeGb = NULL,
  sourceProvenanceHash = NULL, machineType = NULL)
}
\arguments{
\item{logStreamingOption}{Option to define build log streaming behavior to Google Cloud}

\item{requestedVerifyOption}{Requested verifiability options}

\item{substitutionOption}{Option to specify behavior when there is an error in the substitution}

\item{diskSizeGb}{Requested disk size for the VM that runs the build}

\item{sourceProvenanceHash}{Requested hash for SourceProvenance}

\item{machineType}{Compute Engine machine type on which to run the build}
}
\value{
BuildOptions object
}
\description{
BuildOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional arguments to enable specific features of builds.
}
