% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildStep}
\alias{BuildStep}
\title{BuildStep Object}
\usage{
BuildStep(id = NULL, secretEnv = NULL, dir = NULL, timing = NULL,
  env = NULL, waitFor = NULL, entrypoint = NULL, volumes = NULL,
  args = NULL, timeout = NULL, status = NULL, name = NULL)
}
\arguments{
\item{id}{Unique identifier for this build step, used in `wait_for` to}

\item{secretEnv}{A list of environment variables which are encrypted using a Cloud Key}

\item{dir}{Working directory to use when running this step's container}

\item{timing}{Stores timing information for executing this build step}

\item{env}{A list of environment variable definitions to be used when running a step}

\item{waitFor}{The ID(s) of the step(s) that this build step depends on}

\item{entrypoint}{Entrypoint to be used instead of the build step image's default entrypoint}

\item{volumes}{List of volumes to mount into the build step}

\item{args}{A list of arguments that will be presented to the step when it is started}

\item{timeout}{Time limit for executing this build step}

\item{status}{Status of the build step}

\item{name}{Required}
}
\value{
BuildStep object
}
\description{
BuildStep Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A step in the build pipeline.
}
