% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuiltImage}
\alias{BuiltImage}
\title{BuiltImage Object}
\usage{
BuiltImage(digest = NULL, pushTiming = NULL, name = NULL)
}
\arguments{
\item{digest}{Docker Registry 2}

\item{pushTiming}{Stores timing information for pushing the specified image}

\item{name}{Name used to push the container image to Google Container Registry, as}
}
\value{
BuiltImage object
}
\description{
BuiltImage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An image built by the pipeline.
}
