% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_functions.R
\name{operations.list}
\alias{operations.list}
\title{Lists operations that match the specified filter in the request. If theserver doesn't support this method, it returns `UNIMPLEMENTED`.NOTE: the `name` binding allows API services to override the bindingto use different resource name schemes, such as `users/*/operations`. Tooverride the binding, API services can add a binding such as`'/v1/{name=users/*}/operations'` to their service configuration.For backwards compatibility, the default name includes the operationscollection id, however overriding users must ensure the name bindingis the parent resource, without the operations collection id.}
\usage{
operations.list(name, pageToken = NULL, pageSize = NULL, filter = NULL)
}
\arguments{
\item{name}{The name of the operation's parent resource}

\item{pageToken}{The standard list page token}

\item{pageSize}{The standard list page size}

\item{filter}{The standard list filter}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/container-builder/docs/}{Google Documentation}
}
