% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{RepoSource}
\alias{RepoSource}
\title{RepoSource Object}
\usage{
RepoSource(commitSha = NULL, dir = NULL, branchName = NULL,
  tagName = NULL, projectId = NULL, repoName = NULL)
}
\arguments{
\item{commitSha}{Explicit commit SHA to build}

\item{dir}{Directory, relative to the source root, in which to run the build}

\item{branchName}{Name of the branch to build}

\item{tagName}{Name of the tag to build}

\item{projectId}{ID of the project that owns the Cloud Source Repository}

\item{repoName}{Name of the Cloud Source Repository}
}
\value{
RepoSource object
}
\description{
RepoSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Location of the source in a Google Cloud Source Repository.
}
\seealso{
Other RepoSource functions: \code{\link{projects.triggers.run}}
}
