% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Results}
\alias{Results}
\title{Results Object}
\usage{
Results(buildStepImages = NULL, images = NULL)
}
\arguments{
\item{buildStepImages}{List of build step digests, in the order corresponding to build step}

\item{images}{Container images that were built as a part of the build}
}
\value{
Results object
}
\description{
Results Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Artifacts created by the build pipeline.
}
