% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{RetryBuildRequest}
\alias{RetryBuildRequest}
\title{RetryBuildRequest Object}
\usage{
RetryBuildRequest()
}
\value{
RetryBuildRequest object
}
\description{
RetryBuildRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies a build to retry.
}
\seealso{
Other RetryBuildRequest functions: \code{\link{projects.builds.retry}}
}
