% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Source}
\alias{Source}
\title{Source Object}
\usage{
Source(storageSource = NULL, repoSource = NULL)
}
\arguments{
\item{storageSource}{If provided, get the source from this location in Google Cloud Storage}

\item{repoSource}{If provided, get the source from this location in a Cloud Source}
}
\value{
Source object
}
\description{
Source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Location of the source in a supported storage service.
}
