% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Volume}
\alias{Volume}
\title{Volume Object}
\usage{
Volume(path = NULL, name = NULL)
}
\arguments{
\item{path}{Path at which to mount the volume}

\item{name}{Name of the volume to mount}
}
\value{
Volume object
}
\description{
Volume Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Volume describes a Docker container volume which is mounted into build stepsin order to persist files across build step execution.
}
